test(autofix): move utils coverage tests to companion file
