+++ src/utils/autofix.rs
@@ -55,47 +55,3 @@ pub fn standard_outcome(
         hints,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn dry_run_with_changes_returns_preview_status() {
-        let outcome = standard_outcome(
-            AutofixMode::DryRun,
-            3,
-            Some("homeboy test homeboy --analyze".to_string()),
-            vec!["base hint".to_string()],
-        );
-
-        assert_eq!(outcome.status, "auto_fix_preview");
-        assert!(!outcome.rerun_recommended);
-        assert!(outcome.hints.iter().any(|h| h.contains("Dry-run only")));
-    }
-
-    #[test]
-    fn write_with_changes_recommends_rerun() {
-        let outcome = standard_outcome(
-            AutofixMode::Write,
-            2,
-            Some("homeboy test homeboy --analyze".to_string()),
-            vec![],
-        );
-
-        assert_eq!(outcome.status, "auto_fixed");
-        assert!(outcome.rerun_recommended);
-        assert!(outcome
-            .hints
-            .iter()
-            .any(|h| h.contains("Re-run checks: homeboy test homeboy --analyze")));
-    }
-
-    #[test]
-    fn no_changes_returns_noop() {
-        let outcome = standard_outcome(AutofixMode::Write, 0, None, vec![]);
-
-        assert_eq!(outcome.status, "auto_fix_noop");
-        assert!(!outcome.rerun_recommended);
-    }
-}
